Make Executor an enum of borrowed dyn
This is nice since it avoid allocation (with dyn) and is flexible while
allowing pattern matching.

But it is a bit too flexible. The executor, and even the type of
executor, is dynamic - but it could be known at compile time, and
should never change. We'll have to take solace in the fact that Ops are
like singletons.